requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dotfiles==0.6.4
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matlab==0.1
matplotlib==3.6.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.26.1
packaging==23.0
pandas==1.5.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.2.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.17.0
timm==0.6.12
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
yarl==1.8.2
zipp==3.11.0
